+++ CoinUtils/src/CoinLpIO.cpp
@@ -956,7 +956,7 @@ CoinLpIO::find_obj(FILE *fp) const {
   char buff[1024];
 
   sprintf(buff, "aa");
-  unsigned lbuff = strlen(buff);
+  size_t lbuff = strlen(buff);
 
   while(((lbuff != 8) || (CoinStrNCaseCmp(buff, "minimize", 8) != 0)) &&
 	((lbuff != 3) || (CoinStrNCaseCmp(buff, "min", 3) != 0)) &&
@@ -984,7 +984,7 @@ CoinLpIO::find_obj(FILE *fp) const {
 int
 CoinLpIO::is_subject_to(const char *buff) const {
 
-  unsigned lbuff = strlen(buff);
+  size_t lbuff = strlen(buff);
 
   if(((lbuff == 4) && (CoinStrNCaseCmp(buff, "s.t.", 4) == 0)) ||
      ((lbuff == 3) && (CoinStrNCaseCmp(buff, "st.", 3) == 0)) ||
@@ -1001,7 +1001,7 @@ CoinLpIO::is_subject_to(const char *buff) const {
 int 
 CoinLpIO::first_is_number(const char *buff) const {
 
-  int pos;
+  size_t pos;
   char str_num[] = "1234567890";
 
   pos = strcspn (buff, str_num);
@@ -1015,7 +1015,7 @@ CoinLpIO::first_is_number(const char *buff) const {
 int 
 CoinLpIO::is_sense(const char *buff) const {
 
-  int pos;
+  size_t pos;
   char str_sense[] = "<>=";
 
   pos = strcspn(buff, str_sense);
@@ -1039,7 +1039,7 @@ CoinLpIO::is_sense(const char *buff) const {
 int 
 CoinLpIO::is_free(const char *buff) const {
 
-  unsigned lbuff = strlen(buff);
+  size_t lbuff = strlen(buff);
 
   if((lbuff == 4) && (CoinStrNCaseCmp(buff, "free", 4) == 0)) {
     return(1);
@@ -1051,7 +1051,7 @@ CoinLpIO::is_free(const char *buff) const {
 int 
 CoinLpIO::is_inf(const char *buff) const {
 
-  unsigned lbuff = strlen(buff);
+  size_t lbuff = strlen(buff);
 
   if((lbuff == 3) && (CoinStrNCaseCmp(buff, "inf", 3) == 0)) {
     return(1);
@@ -1109,7 +1109,7 @@ int
 CoinLpIO::is_invalid_name(const char *name, 
 			  const bool ranged) const {
 
-  unsigned pos, lname, valid_lname = 100;
+  size_t pos, lname, valid_lname = 100;
   char str_valid[] = "1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ\"!#$%&(),.;?@_'`{}~";
 
   if(ranged) {
@@ -1434,7 +1434,7 @@ CoinLpIO::read_row(FILE *fp, char *buff,
 int 
 CoinLpIO::is_keyword(const char *buff) const {
 
-  unsigned lbuff = strlen(buff);
+  size_t lbuff = strlen(buff);
 
   if(((lbuff == 5) && (CoinStrNCaseCmp(buff, "bound", 5) == 0)) ||
      ((lbuff == 6) && (CoinStrNCaseCmp(buff, "bounds", 6) == 0))) {
@@ -1536,7 +1536,7 @@ CoinLpIO::readLp(FILE* fp)
 
   if(read_st == 2) {
     fscanf(fp, "%s", buff);
-    unsigned lbuff = strlen(buff);
+    size_t lbuff = strlen(buff);
 
     if((lbuff != 2) || (CoinStrNCaseCmp(buff, "to", 2) != 0)) {
       char str[8192];
@@ -1995,7 +1995,7 @@ CoinLpIO::print() const {
 // Hash functions slightly modified from CoinMpsIO.cpp
 
 static int
-hash(const char *name, int maxsiz, int length)
+hash(const char *name, int maxsiz, size_t length)
 {
   static int mmult[] = {
     262139, 259459, 256889, 254291, 251701, 249133, 246709, 244247,
@@ -2010,7 +2010,7 @@ hash(const char *name, int maxsiz, int length)
     82183, 79939, 77587, 75307, 72959, 70793, 68447, 66103
   };
   int n = 0;
-  int j;
+  size_t j;
 
   for ( j = 0; j < length; ++j ) {
     int iname = name[j];
@@ -2052,7 +2052,7 @@ CoinLpIO::startHash(char const * const * const names,
   
   for (i=0; i<number; i++) {
     const char *thisName = names[i];
-    int length = strlen(thisName);
+    size_t length = strlen(thisName);
     
     ipos = hash(thisName, maxhash, length);
     if (hashThis[ipos].index == -1) {
@@ -2073,7 +2073,7 @@ CoinLpIO::startHash(char const * const * const names,
   iput = -1;
   for (i=0; i<number; i++) {
     const char *thisName = names[i];
-    int length = strlen(thisName);
+    size_t length = strlen(thisName);
     
     ipos = hash(thisName, maxhash, length);
     
@@ -2188,7 +2188,7 @@ CoinLpIO::findHash(const char *name, int section) const
   if (!maxhash)
     return -1;
 
-  int length = strlen (name);
+  size_t length = strlen (name);
 
   ipos = hash (name, maxhash, length);
   while (1) {
@@ -2230,7 +2230,7 @@ CoinLpIO::insertHash(const char *thisName, int section)
   char **hashNames = names_[section];
 
   int iput = -1;
-  int length = strlen (thisName);
+  size_t length = strlen (thisName);
 
   int ipos = hash (thisName, maxhash, length);
